codebuild-webinar/sample-app/requirements.txt
numpy
requests
twine
pytest-cov